Revert "Gradle update (#637)" and fix dundet jcenter repo deps
This reverts commit 4c49a9cf61705b6f2790e3a59577ef405cba2726.

+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/utils/CpuVideoProcessor.kt
@@ -164,7 +164,7 @@ class CpuVideoProcessor(private val logger: Logger, eglCoreFactory: EglCoreFacto
                 val gValue = rgbaBuffer.data[gLocation].toPositiveInt()
                 val bValue = rgbaBuffer.data[bLocation].toPositiveInt()
 
-                val newValue = ((rValue + gValue + bValue) / (3.0)).toInt().toByte()
+                val newValue = ((rValue + gValue + bValue) / (3.0)).toByte()
 
                 rgbaBuffer.data.put(rLocation, newValue)
                 rgbaBuffer.data.put(gLocation, newValue)
